Restructure CLI into per-command classes; remove admin backfill UI
Mirrors the ActivityPub plugin's CLI layout:

- `Atmosphere\Cli` is a thin registry (`Cli::register()` called from
  `atmosphere.php` only under WP_CLI).
- `Atmosphere\Cli\Command` is the base class with a `version`
  subcommand (`wp atmosphere version`).
- `Atmosphere\Cli\Backfill_Command` extends `WP_CLI_Command` and
  carries the backfill logic in `__invoke()`. `parse_ids()` is the
  only static helper exposed for unit testing.

Tests follow the same shape — `tests/phpunit/tests/cli/` mirrors
`includes/cli/`. The `parse_ids` data-provider tests move out of
`class-test-backfill.php` (which now covers only the
`Backfill::get_unsynced_post_ids` helper) and into the new
`Test_Backfill_Command` class. A stub for `\WP_CLI_Command` lives in
`tests/phpunit/class-wp-cli-command.php` so the autoloader can
resolve the command class without the WP-CLI runtime being present.

Admin backfill UI removed:

- `Backfill::register/handle_count/handle_batch` and the
  `wp_ajax_atmosphere_backfill_*` actions are gone.
- The "Start Backfill" settings field, its render method, the
  backfill JS (`assets/js/settings.js`, deleted), the backfill CSS
  block, and the localized nonce are all removed.
- `Backfill` is now a query helper with one public static
  (`get_unsynced_post_ids`) used by the CLI.

Operators who want to bulk-sync run `wp atmosphere backfill`.
